django/requirements.txt
bokeh==2.1.1
dask[complete]==2.19.0
distributed==2.19.0
django==3.0.7
djangorestframework==3.11.0
django-anymail==6.1.0
django-picklefield==3.0.1
requests==2.24.0
